Split variable_operations.h
