+++ core/variable_logical.cpp
