+++ core/variable_reduction.cpp
