+++ core/variable_trigonometry.cpp
